+++ provider-hedera/src/test/java/network/arkane/provider/hedera/bridge/HederaTransactionGatewayTest.java
@@ -18,7 +18,7 @@ import network.arkane.provider.sign.domain.TransactionSignature;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.TimeoutException;
@@ -42,7 +42,7 @@ class HederaTransactionGatewayTest {
         TransactionSignature signature = signer.createSignature(HbarTransferSignable.builder()
                                                                                     .from(HederaTestFixtures.getAccountId().toString())
                                                                                     .to("0.0.2256926")
-                                                                                    .amount(new BigDecimal("1000"))
+                                                                                    .amount(new BigInteger("1000"))
                                                                                     .build(),
                                                                 HederaSecretKey.builder().key(HederaTestFixtures.getOperatorKey()).build());
 
@@ -56,7 +56,7 @@ class HederaTransactionGatewayTest {
         TransactionSignature signature = signer.createSignature(HbarTransferSignable.builder()
                                                                                     .from("0.0.2256926")
                                                                                     .to("0.0.2151494")
-                                                                                    .amount(new BigDecimal("0.01"))
+                                                                                    .amount(new BigInteger("0.01"))
                                                                                     .build(),
                                                                 HederaSecretKey.builder()
                                                                                .key(PrivateKey.fromString(
